Feature/create threads

+++ src/deleteFile.go
@@ -2,16 +2,22 @@ package main
 
 import (
 	"fmt"
-	"github.com/julienschmidt/httprouter"
 	"log"
 	"net/http"
 	"os"
 	"path/filepath"
+	"strings"
+	"sync"
 )
 
-func DeleteFile(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
+var mutexDelete sync.Mutex
 
-	fileName := ps.ByName("filename")
+func DeleteFile(w http.ResponseWriter, r *http.Request) {
+	urlParts := strings.Split(r.URL.Path, "/")
+	fileName := urlParts[len(urlParts)-1]
+
+	DeleteInProgress.Store(0, true)
+
 	log.Println("[DELETE] /delete/" + fileName)
 
 	if r.Method != http.MethodDelete {
@@ -24,19 +30,30 @@ func DeleteFile(w http.ResponseWriter, r *http.Request, ps httprouter.Params) {
 		return
 	}
 
-	filePath := filepath.Join(UploadDir, fileName)
-	_, err := os.Stat(filePath)
+	filePath := filepath.Join(BaseDir, fileName)
 
+	if _, ok := DeleteInProgress.Load(1); ok {
+		http.Error(w, "Download in progress, cannot delete the file", http.StatusForbidden)
+		return
+	}
+
+	mutexDelete.Lock()
+	_, err := os.Stat(filePath)
 	if err != nil {
+		mutexDelete.Unlock()
 		http.Error(w, fmt.Sprintf("File %s not found", fileName), http.StatusNotFound)
 		return
 	}
 
 	err = os.Remove(filePath)
 	if err != nil {
+		mutexDelete.Unlock()
 		http.Error(w, fmt.Sprintf("Error deleting the file: %v", err), http.StatusInternalServerError)
 		return
 	}
+
+	DeleteInProgress.Store(0, false)
+	mutexDelete.Unlock()
 
 	w.WriteHeader(http.StatusOK)
 	w.Write([]byte("File deleted successfully!\n"))
